+++ src/components/Community.js
@@ -2,7 +2,7 @@ import React from "react";
 
 const Community = () => {
     return(
-        <section id="apropos">
+        <section id="community">
             <div className="discord">
                     <iframe src="https://discord.com/widget?id=817218229377957938&theme=dark" width="350" height="500" allowtransparency="true" frameBorder="0" sandbox="allow-popups allow-popups-to-escape-sandbox allow-same-origin allow-scripts" title="discord" className="card"/>
                 </div>
@@ -12,6 +12,7 @@ const Community = () => {
                     <li className="card"><a href="https://www.linkedin.com/company/wordlet/">LinkedIn</a></li>
                     <li className="card"><a href="https://www.facebook.com/WordletTheWordWallet">Facebook</a></li>
                 </ul>
+                <img src="lava1.jpg" id="mountains_front" alt=""/>
         </section>
     )
 }
