--- src/components/Community/Contacts.js
@@ -1,14 +1,0 @@
-import React from "react"
-
-const Contacts = () => {
-    return (
-        <ul id="contacts">
-            <li class="card"><a href="mailto:wordletthewordwallet@gmail.com" >mail</a></li>
-            <li class="card"><a href="https://discord.gg/zmRRJcVJzM">Discord</a></li>
-            <li class="card"><a href="https://www.linkedin.com/company/wordlet/">LinkedIn</a></li>
-            <li class="card"><a href="https://www.facebook.com/WordletTheWordWallet">Facebook</a></li>
-        </ul>
-    )
-}
-
-export default Contacts
