--- src/components/Community/Discord.js
@@ -1,11 +1,0 @@
-import React from "react"
-
-const Discord = () => {
-return(
-<div className="discord">
-<iframe src="https://discord.com/widget?id=817218229377957938&theme=dark" width="350" height="500" allowtransparency="true" frameBorder="0" sandbox="allow-popups allow-popups-to-escape-sandbox allow-same-origin allow-scripts" title="discord" className="card"/>
-</div>
-)
-}
-
-export default Discord
