--- src/components/UserData/UserData.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-import UserTokens from "./UserTokens"
-
-const UserData = () => {
-    return(
-        <UserTokens/>
-    )
-}
-
-export default UserData
